Add functions and make When a proxy for them for BC
