+++ composer.json
@@ -11,7 +11,8 @@
     "autoload": {
         "psr-0": {
             "React\\Promise": "src/"
-        }
+        },
+        "files": ["src/React/Promise/functions.php"]
     },
     "extra": {
         "branch-alias": {
